+++ index.ts
@@ -1,3 +1,4 @@
+import {read_ir, opc, opd, run, brun} from "./clvm_tools/cmds";
 export * from "./clvm_tools/binutils";
 export * from "./clvm_tools/clvmc";
 export * from "./clvm_tools/cmds";
@@ -9,6 +10,34 @@ export * from "./clvm_tools/sha256tree";
 
 import {setStdout, TPrinter} from "./platform/print";
 import {initialize as initClvm} from "clvm";
+
+const COMMANDS: Record<string, (args: string[]) => unknown> = {
+  read_ir,
+  opc,
+  opd,
+  run,
+  brun,
+};
+
+/**
+ * Dispatch cli command.
+ * - `go("run", "(mod ARGUMENT (+ ARGUMENT 3))")`
+ * - `go("brun", "(+ 1 (q . 3))", "--time")`
+ * 
+ * @param {...string[]} args 
+ */
+export function go(...args: string[]){
+  if(!args || args.length < 1){
+    throw new Error("You need specify command");
+  }
+  const commandName = args[0];
+  const command = COMMANDS[commandName];
+  if(!command){
+    throw new Error(`Unknown command: ${commandName}`);
+  }
+  
+  return command(args);
+}
 
 /**
  * Change print function. Default is `console.log`.